# Speed up history operations