reduces time to complete operations (add/undo/redo) by round 40%

quick benchmarking with 100k rows (x 50 Columns)
before
undo: 2924.723876953125 ms
undo: 2517.576904296875 ms
undo: 2860.81005859375 ms
undo: 2541.362060546875 ms
redo: 2683.650146484375 ms
redo: 2658.299072265625 ms
redo: 2681.257080078125 ms
redo: 2728.52099609375 ms

after
undo: 1793.409912109375 ms
undo: 1363.796875 ms
undo: 1396.85791015625 ms
undo: 1488.695068359375 ms
redo: 1727.51513671875 ms
redo: 1397.403076171875 ms
redo: 1425.426025390625 ms
redo: 1458.501953125 ms